Follo reviews and put both classic and net in the same file with an prepro directive.
