+++ tests/bgen/bgen-tests.csproj
@@ -29,14 +29,11 @@
     <Compile Include="..\generator\BGenTool.cs">
       <Link>BGenTool.cs</Link>
     </Compile>
-    <Compile Include="..\..\src\bgen\AttributeFactory.Classic.cs">
-      <Link>AttributeFactory.Classic.cs</Link>
+    <Compile Include="..\..\src\bgen\AttributeFactory.ConstructorArguments.cs">
+      <Link>AttributeFactory.ConstructorArguments.cs</Link>
     </Compile>
     <Compile Include="..\..\src\bgen\AttributeFactory.cs">
       <Link>AttributeFactory.cs</Link>
-    </Compile>
-    <Compile Include="..\..\src\bgen\AttributeFactory.NET.cs">
-      <Link>AttributeFactory.NET.cs</Link>
     </Compile>
     <Compile Include="..\generator\AttributeFactoryTests.cs">
       <Link>AttributeFactoryTests.cs</Link>
