+++ tests/generator/generator-tests.csproj
@@ -38,14 +38,11 @@
     <PackageReference Include="MSBuild.StructuredLogger" Version="2.1.758" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\..\src\bgen\AttributeFactory.Classic.cs">
-      <Link>AttributeFactory.Classic.cs</Link>
+    <Compile Include="..\..\src\bgen\AttributeFactory.ConstructorArguments.cs">
+      <Link>AttributeFactory.ConstructorArguments.cs</Link>
     </Compile>
     <Compile Include="..\..\src\bgen\AttributeFactory.cs">
       <Link>AttributeFactory.cs</Link>
-    </Compile>
-    <Compile Include="..\..\src\bgen\AttributeFactory.NET.cs">
-      <Link>AttributeFactory.NET.cs</Link>
     </Compile>
     <Compile Include="..\..\tools\common\SdkVersions.cs">
       <Link>SdkVersions.cs</Link>
